requirements.txt
certifi==2021.5.30
cftime==1.5.0
charset-normalizer==2.0.4
idna==3.2
netCDF4==1.5.7
numpy==1.21.1
python-dateutil==2.8.2
python-dotenv==0.19.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
